refact: add types.ts for products.ts
