+++ online-store/src/data/products.ts
@@ -1,15 +1,4 @@
-export type ProductType = {
-    id: string;
-    title: string;
-    model: string;
-    manufacturer: string;
-    year: string;
-    price: string;
-    memory: string;
-    quantity: number;
-    color: string;
-    image: string;
-};
+import { ProductType } from "./types";
 
 export const products: ProductType[] = [
     {
