+++ online-store/src/data/types.ts
@@ -1,0 +1,12 @@
+export type ProductType = {
+    id: string;
+    title: string;
+    model: string;
+    manufacturer: string;
+    year: string;
+    price: string;
+    memory: string;
+    quantity: number;
+    color: string;
+    image: string;
+};
